chore: rename claude_agent_service.py to specialized_agent_service.py (#1819)
Align filename with the class name (SpecializedAgentService) that was
renamed during #1794. Update three inline imports in agent_config.py.
