+++ autobot-backend/services/specialized_agent_service.py
